Refactor typename lookup

+++ archetypemodel/src/main/java/com/nedap/archie/rm/datavalues/quantity/DvCount.java
@@ -3,5 +3,5 @@ package com.nedap.archie.rm.datavalues.quantity;
 /**
  * Created by pieter.bos on 04/11/15.
  */
-public class DvCount extends DvAmount<Integer> {
+public class DvCount extends DvAmount<Long> {
 }
